# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_cocursor_backend_internal_application_cursor.GenerationWithSource:
    properties:
      generationUUID:
        description: 生成的唯一标识符
        type: string
      source:
        description: 工作区 ID
        type: string
      textDescription:
        description: AI 的实际回复内容或操作描述
        type: string
      type:
        description: 生成类型：composer/chat/tab
        type: string
      unixMs:
        description: Unix 时间戳（毫秒）
        type: integer
    type: object
...
  github_com_cocursor_backend_internal_application_cursor.PromptWithSource:
    properties:
      prompt:
        additionalProperties: true
        type: object
      source:
        description: 工作区 ID
        type: string
    type: object
  github_com_cocursor_backend_internal_application_cursor.SessionWithSource:
    properties:
      composerId:
        description: 会话唯一标识符
        type: string
      contextUsagePercent:
        description: 代码变更统计（关键字段）
        type: number
      createdAt:
        description: 创建时间戳（毫秒）
        type: integer
      createdOnBranch:
        description: 创建时的 Git 分支
        type: string
      filesChangedCount:
        description: 变更的文件数量
        type: integer
      forceMode:
        description: 强制模式
        type: string
      hasBlockingPendingActions:
        type: boolean
      hasUnreadMessages:
        description: 状态标志
        type: boolean
      isArchived:
        type: boolean
      isDraft:
        type: boolean
      isSpec:
        type: boolean
      isWorktree:
        type: boolean
      lastUpdatedAt:
        description: 最后更新时间戳（毫秒）
        type: integer
      name:
        description: 会话名称
        type: string
      numSubComposers:
        description: 层级关系
        type: integer
      referencedPlans:
        description: 引用的计划列表
        items:
          type: string
        type: array
      source:
        description: 工作区 ID
        type: string
      subtitle:
        description: 涉及的文件列表（逗号分隔）
        type: string
      totalLinesAdded:
        description: 总添加行数
        type: integer
      totalLinesRemoved:
        description: 总删除行数
        type: integer
      type:
        description: 基础信息
        type: string
      unifiedMode:
        description: 模式信息
        type: string
    type: object
...
  github_com_cocursor_backend_internal_domain_cursor.DailyAcceptanceStats:
    properties:
      composer_acceptance_rate:
        description: Composer 接受率（百分比，计算字段）
        type: number
      composerAcceptedLines:
        description: Composer 接受的代码行数
        type: integer
      composerSuggestedLines:
        description: Composer 建议的代码行数
        type: integer
      data_quality:
        description: 数据质量标识：normal/warning/invalid
        type: string
      date:
        description: 日期 YYYY-MM-DD
        type: string
      tab_acceptance_rate:
        description: Tab 接受率（百分比，计算字段）
        type: number
      tabAcceptedLines:
        description: Tab 接受的代码行数
        type: integer
      tabSuggestedLines:
        description: Tab 建议的代码行数
        type: integer
      warning_message:
        description: 警告信息
        type: string
    type: object
...
  github_com_cocursor_backend_internal_domain_cursor.ProjectInfo:
    properties:
      created_at:
        description: 项目首次发现时间
        type: string
      git_branch:
        description: Git 分支（如果有）
        type: string
      git_remote_url:
        description: Git 远程仓库 URL（如果有）
        type: string
      last_updated_at:
        description: 最后更新时间
        type: string
      project_id:
        description: 项目唯一 ID
        type: string
      project_name:
        description: 项目名称（唯一）
        type: string
      workspaces:
        description: 包含的所有工作区
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.WorkspaceInfo'
        type: array
    type: object
...
  github_com_cocursor_backend_internal_domain_cursor.WorkspaceInfo:
    properties:
      git_branch:
        description: Git 分支
        type: string
      git_remote_url:
        description: Git 远程 URL
        type: string
      is_active:
        description: 是否为当前活跃的工作区
        type: boolean
      is_primary:
        description: 是否为主工作区（最新的）
        type: boolean
      path:
        description: 项目路径
        type: string
      project_name:
        description: 所属项目名
        type: string
      workspace_id:
        description: Cursor 工作区 ID
        type: string
    type: object
  github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse:
    properties:
      code:
        type: integer
      detail:
        type: string
      message:
        type: string
    type: object
  github_com_cocursor_backend_internal_interfaces_http_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  internal_interfaces_http_handler.ActivateProjectRequest:
    properties:
      path:
        description: 项目路径
        type: string
      timestamp:
        description: 时间戳
        type: integer
    required:
    - path
    type: object
  internal_interfaces_http_handler.ActivateProjectResponse:
    properties:
      is_active:
        type: boolean
      message:
        type: string
      project_id:
        type: string
      project_name:
        type: string
      success:
        type: boolean
      workspace_info:
        $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.WorkspaceInfo'
    type: object
  internal_interfaces_http_handler.ProjectAcceptanceStatsResponse:
    properties:
      merged_stats:
        $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.DailyAcceptanceStats'
      project_name:
        type: string
      raw_stats:
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.DailyAcceptanceStats'
        type: array
      workspaces:
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.WorkspaceInfo'
        type: array
    type: object
  internal_interfaces_http_handler.ProjectGenerationsResponse:
    properties:
      generations:
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_application_cursor.GenerationWithSource'
        type: array
      project_name:
        type: string
      total:
        type: integer
      workspaces:
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.WorkspaceInfo'
        type: array
    type: object
  internal_interfaces_http_handler.ProjectPromptsResponse:
    properties:
      project_name:
        type: string
      prompts:
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_application_cursor.PromptWithSource'
        type: array
      total:
        type: integer
      workspaces:
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.WorkspaceInfo'
        type: array
    type: object
  internal_interfaces_http_handler.ProjectSessionsResponse:
    properties:
      project_name:
        type: string
      sessions:
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_application_cursor.SessionWithSource'
        type: array
      total:
        type: integer
      workspaces:
        items:
          $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.WorkspaceInfo'
        type: array
    type: object
host: localhost:19960
info:
  contact: {}
  description: cocursor 守护进程 API 服务
  title: cocursor Daemon API
  version: "1.0"
paths:
...
  /project/{project_name}/generations:
    get:
      consumes:
      - application/json
      parameters:
      - description: 项目名称
        in: path
        name: project_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_interfaces_http_handler.ProjectGenerationsResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
      summary: 查询项目的 AI 生成记录
      tags:
      - 项目
  /project/{project_name}/prompts:
    get:
      consumes:
      - application/json
      parameters:
      - description: 项目名称
        in: path
        name: project_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_interfaces_http_handler.ProjectPromptsResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
      summary: 查询项目的 AI 对话历史
      tags:
      - 项目
  /project/{project_name}/sessions:
    get:
      consumes:
      - application/json
      parameters:
      - description: 项目名称
        in: path
        name: project_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_interfaces_http_handler.ProjectSessionsResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
      summary: 查询项目的 Composer 会话
      tags:
      - 项目
  /project/{project_name}/stats/acceptance:
    get:
      consumes:
      - application/json
      parameters:
      - description: 项目名称
        in: path
        name: project_name
        required: true
        type: string
      - description: 开始日期 YYYY-MM-DD，默认最近 7 天
        in: query
        name: start_date
        type: string
      - description: 结束日期 YYYY-MM-DD，默认今天
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_interfaces_http_handler.ProjectAcceptanceStatsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
      summary: 查询项目的接受率统计
      tags:
      - 项目
  /project/activate:
    post:
      consumes:
      - application/json
      description: 接收前端上报的当前项目路径，更新活跃状态
      parameters:
      - description: 激活请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.ActivateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_interfaces_http_handler.ActivateProjectResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
      summary: 激活项目
      tags:
      - 项目
  /project/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_cocursor_backend_internal_domain_cursor.ProjectInfo'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cocursor_backend_internal_interfaces_http_response.ErrorResponse'
      summary: 列出所有已发现的项目
      tags:
      - 项目
...
schemes:
- http
swagger: "2.0"